+++ plugins/plugin-github/src/utils/constants.ts
@@ -1,0 +1,1 @@
+export const defaultDescription = 'Where the world builds software.';